// peso_lucky_admin-main/src/pages/Record/Test2/data.d.ts
export interface Params {}
export interface ListItem {
  game_code_type: string;
  award_amount: string;
  second_draw_result: any;
  code_type: string;
  issue_number: any;
  seq: number;
  game_code: string;
  operator: string;
  created_at: number;
  updated_at: number;
  canceled_at: number;
  open_time: number;
  close_time: number;
  end_time: number;
  round_type: number;
  prize_content: string;
  second_content: string;
  status: number;
  next_seq: string;
  cancel_operator: string;
  remark: string;
  seq: string;
  game_type: number;
  draw_time: number;
  second_draw_time: number;
  prize_time: number;
}